fix(color-mode): resolve ESLint issues in color-mode.tsx
