+++ src/components/ui/color-mode.tsx
@@ -1,47 +1,26 @@
 "use client";
 
+import { useColorMode } from "@/hooks/color-mode.hooks";
 import type { IconButtonProps } from "@chakra-ui/react";
 import { ClientOnly, IconButton, Skeleton } from "@chakra-ui/react";
-import { ThemeProvider, useTheme } from "next-themes";
+import { ThemeProvider } from "next-themes";
 import type { ThemeProviderProps } from "next-themes";
 import * as React from "react";
 import { LuMoon, LuSun } from "react-icons/lu";
 
-export interface ColorModeProviderProps extends ThemeProviderProps {}
-
-export function ColorModeProvider(props: ColorModeProviderProps) {
+// 移除空介面，直接使用 ThemeProviderProps
+export function ColorModeProvider(props: ThemeProviderProps) {
     return (
         <ThemeProvider attribute="class" disableTransitionOnChange {...props} />
     );
 }
 
-export function useColorMode() {
-    const { resolvedTheme, setTheme } = useTheme();
-    const toggleColorMode = () => {
-        setTheme(resolvedTheme === "light" ? "dark" : "light");
-    };
-    return {
-        colorMode: resolvedTheme,
-        setColorMode: setTheme,
-        toggleColorMode,
-    };
-}
-
-export function useColorModeValue<T>(light: T, dark: T) {
-    const { colorMode } = useColorMode();
-    return colorMode === "light" ? light : dark;
-}
-
-export function ColorModeIcon() {
-    const { colorMode } = useColorMode();
-    return colorMode === "light" ? <LuSun /> : <LuMoon />;
-}
-
-interface ColorModeButtonProps extends Omit<IconButtonProps, "aria-label"> {}
+// 將 hooks 和 utility functions 移至單獨的文件 color-mode.hooks.ts
+type ColorModeButtonProps = Omit<IconButtonProps, "aria-label">;
 
 export const ColorModeButton = React.forwardRef<
-  HTMLButtonElement,
-  ColorModeButtonProps
+    HTMLButtonElement,
+    ColorModeButtonProps
 >(function ColorModeButton(props, ref) {
     const { toggleColorMode } = useColorMode();
     return (
@@ -65,3 +44,9 @@ export const ColorModeButton = React.forwardRef<
         </ClientOnly>
     );
 });
+
+// 將 Icon 組件保留在這個文件中
+export function ColorModeIcon() {
+    const { colorMode } = useColorMode();
+    return colorMode === "light" ? <LuSun /> : <LuMoon />;
+}
