+++ src/hooks/color-mode.hooks.ts
@@ -1,0 +1,18 @@
+import { useTheme } from "next-themes";
+
+export function useColorMode() {
+    const { resolvedTheme, setTheme } = useTheme();
+    const toggleColorMode = () => {
+        setTheme(resolvedTheme === "light" ? "dark" : "light");
+    };
+    return {
+        colorMode: resolvedTheme,
+        setColorMode: setTheme,
+        toggleColorMode,
+    };
+}
+
+export function useColorModeValue<T>(light: T, dark: T) {
+    const { colorMode } = useColorMode();
+    return colorMode === "light" ? light : dark;
+}
